KARTIK_ISTP/Kartik_work/requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# Utils
tqdm>=4.65.0
pyyaml>=6.0
kaggle>=1.5.16
yt-dlp>=2023.10.0

# Optional but recommended
wandb>=0.15.0  # For experiment tracking
